requirements.txt
deepsport-utilities >= 4.0.9
torch
torchvision
tqdm
yacs
pytorch-ignite
